Checkin Jooq Changes
This makes life easeir for bootstrapping because we have the migrations
in the same module. To migrate you have to compile, but if compilation
depends on jooq and jooq depends on the db being up and fully migrated
you are stuck. With this change updating jooq is manual, but easy
process ./mvwn generateJooq. And we simply checkin the source