pointed model to remote ollama and experimented with context size
I've been running the backend and ollama on the same local machine. In
this commit I've configured it so that the backend server can run my
local dev box while the ollam stack remains on more capable machine
with an RTX card.

Obivously this should all be environmentally configured. I'm building
towards some exploration of running parts of this in the cloud so I'll
continue to iterate.