original non-streaming working again
I had to make quite a few changes following upgrades

The underlying SDK seems to now be splitting out thinking tokens.

Rather than relying on content I had to fall to a lower level
modelResponse object which allows you to access the thinking tokens
as well as the content contents.

I also discovered this object gives us access to information like tokens
used. I've started logging this data, in the future we should pass that
back to the client